# Optimize campaigns and campaign endpoint