--- a/backend/app/methods/campaignCallMethods.js
+++ b/backend/app/methods/campaignCallMethods.js
@@ -1,10 +1,9 @@
 const { CampaignCall } = require('../models')
 const { getPopulatedCampaignCallObject } = require('../utilities/campaignCalls')
 
-function getCampaignCallDetail(req, res) {
-  return getPopulatedCampaignCallObject(req.params.id)
-    .then(campaignCallObject => res.json(campaignCallObject))
-    .catch(err => res.send(err))
+async function getCampaignCallDetail(req, res) {
+  const campaignCall = await getPopulatedCampaignCallObject(req.params.id)
+  return res.json(campaignCall)
 }
 
 function getCampaignCall(req, res) {