--- a/backend/app/methods/campaignMethods.js
+++ b/backend/app/methods/campaignMethods.js
@@ -69,10 +69,7 @@ exports.getCampaigns = async function(req, res) {
   Campaign
     .find({bot: bot})
     .populate({
-      path: 'campaignActions',
-      populate: {
-        path: 'userConversations'
-      }
+      path: 'campaignActions'
     })
     .exec(function(err, campaigns) {
       if (err) return res.send(err)