--- a/backend/app/methods/campaignMethods.js
+++ b/backend/app/methods/campaignMethods.js
@@ -51,9 +51,6 @@ exports.getCampaign = function(req, res) {
     .findOne({ _id: req.params.id, bot: bot })
     .populate({
       path: 'campaignActions',
-      populate: {
-        path: 'userConversations'
-      }
     })
     .exec(function(err, campaign) {
       if (err) return res.send(err)