For some reason we were populating userConversation on each of these endpoints
when they were not needed  

By removing these populate it took response times down from 
17 seconds to 500ms for campaigns
5 seconds to 320ms for campaign detail

The one page that is still sometimes slow is campaignAction detail
to improve the speed of this page, we should make the userConversation table into a paginated table with infinite scroll
--> I created a github issue here: https://github.com/CallParty/CallParty/issues/360

resolves: https://github.com/CallParty/CallParty/issues/358