@mhfowler 2017-09-11
Also added a loader to campaignDetail page